Finished Project, just need comments
Linked in both audio clips and added in controls for the user to be able to use. Also embedded in the audio clips so that if the user's browser doesn't support HTML5 audio, they would still be able to listen. Finally, I added in styles that would affect the look of the audio clips by changing their color, adding in a box shadow, etc.